+++ src/telliot_feed_examples/cli.py
@@ -3,17 +3,14 @@
 A simple interface for interacting with telliot example feed functionality.
 """
 import asyncio
-from typing import Tuple
 
 import click
 from click.core import Context
 from telliot_core.apps.telliot_config import TelliotConfig
-from telliot_core.contract.contract import Contract
-from telliot_core.directory.tellorx import tellor_directory
-from telliot_core.model.endpoints import RPCEndpoint
 
 from telliot_feed_examples.feeds import LEGACY_DATAFEEDS
 from telliot_feed_examples.reporters.interval import IntervalReporter
+from telliot_feed_examples.utils.contract import get_tellor_contracts
 from telliot_feed_examples.utils.log import get_logger
 from telliot_feed_examples.utils.oracle_write import tip_query
 
@@ -23,34 +20,6 @@ logger = get_logger(__name__)
 
 # Get default configs from ~/telliot/
 cfg = TelliotConfig()
-
-
-def get_tellor_contracts(
-    private_key: str, chain_id: int, endpoint: RPCEndpoint
-) -> Tuple[Contract, Contract]:
-    """Get Contract objects per telliot configuration and
-    CLI flag options."""
-    endpoint.connect()
-
-    tellor_oracle = tellor_directory.find(chain_id=chain_id, name="oracle")[0]
-    oracle = Contract(
-        address=tellor_oracle.address,
-        abi=tellor_oracle.abi,
-        node=endpoint,
-        private_key=private_key,
-    )
-    oracle.connect()
-
-    tellor_master = tellor_directory.find(chain_id=chain_id, name="master")[0]
-    master = Contract(
-        address=tellor_master.address,
-        abi=tellor_master.abi,
-        node=endpoint,
-        private_key=private_key,
-    )
-    master.connect()
-
-    return master, oracle
 
 
 # Main CLI options
